make_gaussian: Move to eradiate.spectral.response
