+++ docs/rst/reference_api/srf_tools.rst
@@ -5,11 +5,6 @@
 
 .. py:currentmodule:: eradiate.srf_tools
 
-Generators
-----------
-
-.. autofunction:: make_gaussian
-
 Trimming algorithms
 -------------------
